+++ CHANGES.rst
@@ -5,6 +5,10 @@ v4.6.2
 
 Usability, bells and whistles
 -----------------------------
+* The Fetch, Push and Pull dialogs now display the git commands that will be run.
+  The Pull dialog now selects the remote branch automatically.
+  (`#729 <https://github.com/git-cola/git-cola/issues/729>`_)
+
 * The `cola.refreshonfocus` and `cola.inotify` configuration settings are now accessible
   from the `git cola config` settings dialog.